requirements.txt
pyshark 
click 
pyyaml